+++ pyincore/analyses/bridgedamage/bridgedamage.py
@@ -9,6 +9,7 @@ import collections
 import concurrent.futures
 import random
 from itertools import repeat
+import copy
 
 from pyincore import AnalysisUtil, GeoUtil
 from pyincore import BaseAnalysis, HazardService, FragilityService
@@ -126,7 +127,6 @@ class BridgeDamage(BaseAnalysis):
             fragility_set = None
             if bridge["id"] in fragility_sets:
                 fragility_set = fragility_sets[bridge["id"]]
-
             result.append(self.bridge_damage_analysis(bridge, fragility_set,
                                                       hazard_type,
                                                       hazard_dataset_id,
@@ -195,19 +195,21 @@ class BridgeDamage(BaseAnalysis):
 
             hazard_val = hazard_resp[0]['hazardValue']
             hazard_std_dev = 0.0
-            adjusted_fragility_set = fragility_set
+            adjusted_fragility_set = copy.deepcopy(fragility_set)
 
             # TODO Get this from API once implemented
             if use_hazard_uncertainty:
                 hazard_std_dev = random.random()
 
             if use_liquefaction and 'liq' in bridge['properties']:
-                for fragility in fragility_set["fragilityCurves"]:
-                    adjusted_fragility_set.append(
-                        AnalysisUtil.adjust_fragility_for_liquefaction(
-                            fragility, bridge['properties']['liq']))
+                for fragility in adjusted_fragility_set["fragilityCurves"]:
+                    AnalysisUtil.adjust_fragility_for_liquefaction(
+                            fragility, bridge['properties']['liq'])
 
-            dmg_probability = AnalysisUtil.calculate_limit_state(fragility_set, hazard_val, std_dev=hazard_std_dev)
+            dmg_probability = AnalysisUtil.calculate_limit_state(adjusted_fragility_set,
+                                                                 hazard_val,
+                                                                 std_dev=hazard_std_dev)
+
             retrofit_cost = BridgeUtil.get_retrofit_cost(fragility_key)
             retrofit_type = BridgeUtil.get_retrofit_type(fragility_key)